Harden MVCC EBR lifecycle, clippy-clean core/VDBE/pager/btree, and sweep harness quality gates
MVCC (fsqlite-mvcc):
- ebr: add nested_version_guards_pin_and_unpin_independently test proving
  inner/outer guard independence and correct metric accounting
- lifecycle: add test_version_guard_deferred_value_freed_after_abort verifying
  aborted txn retirements are reclaimed after unpin + epoch advance; add
  test_concurrent_same_page_writers_preserve_all_committed_versions proving
  8-thread same-page contention retains full version chain
- invariants: fix TxnManager::alloc_commit_seq to avoid holding lock across
  fetch_add; add explicit drop(active) in finish_commit_seq to release lock
  before stable_commit_seq update; enrich Debug impls with active_commits,
  stable_commit_seq, and guard_registry fields
- core_types: replace manual generation-mismatch panic with assert! macro;
  add #[allow(clippy::too_many_lines)] on try_cleanup_orphaned_slot
- gc: relax EBR metric assertions from >= N+1 to > N for robustness

Core (fsqlite-core):
- connection: remove unnecessary .clone() on update_cols (TriggerEvent::Update);
  add #[allow(clippy::unused_self)] on map_insert_source_row_to_table_row
- por: wrap truncation cast in #[allow(clippy::cast_possible_truncation)] block
- repair_engine: eliminate redundant data.clone() before corruption; use
  f64::from() instead of `as f64` casts for lossless numeric conversion

B-tree (fsqlite-btree):
- btree_invariant_tests: use HashSet::insert() return value instead of
  separate contains+insert; suppress unused Result from cache.insert() with
  let _ binding across 4 test sites

Pager (fsqlite-pager):
- arc_cache: enrich ArcCacheInner Debug impl with hits, misses, ghost_hits_b1,
  ghost_hits_b2, evictions_t1, evictions_t2, version_coalesce_count, admits
- pager: add explicit drop(inner) after lock guard usage in 3 test sites;
  add #[allow(clippy::similar_names, clippy::cast_possible_truncation)] on
  test_cache_eviction_under_pressure

VDBE (fsqlite-vdbe):
- vectorized_dispatch: replace &[tasks.clone()] with std::slice::from_ref;
  use map_or_else instead of map+unwrap_or_else for artifact path; add
  #[allow(clippy::too_many_lines)] on morsel_dispatch_e2e_replay_emits_artifact
- vectorized_ops: add #[allow(clippy::cast_possible_truncation)] on 2 filter
  tests; replace "delta".as_bytes() with b"delta"; eliminate redundant .clone()
  on batch before apply_selection
- vectorized_prop_tests: eliminate 5 redundant .clone() calls on values/sorted
  vectors; replace nested match-continue with let-else patterns for Option
  destructuring; use iter().flatten() instead of filter_map for semi-join keys;
  replace match-Err-return with let-else for agg_result

Types (fsqlite-types):
- cx: add #[allow(clippy::redundant_clone)] on test_clone_propagates_trace_ids

Harness (fsqlite-harness, ~40 modules + 12 test files):
- adversarial_search: add cast_possible_truncation allow; simplify
  classify_flip_severity match arms; use mul_add for injection_rate
- backlog_quality_gate: major refactor of gate evaluation and formatting logic
- bin/: clippy fixes across durability_matrix_manifest, perf_baseline_pack_runner,
  scenario_coverage_drift_gate_runner, unit_coverage_drift_gate_runner,
  validation_manifest_runner
- ci_coverage_gate, ci_gate_matrix, confidence_gates: lint suppressions and
  minor refactors
- corpus_ingest, crash_recovery_parity, cross_process_crash_harness: allow
  attrs for truncation casts and similar names
- drift_monitor, e2e_log_schema, e2e_logging_init: clippy fixes for cast
  safety and redundant operations
- execution_waves, extension_parity_matrix, failure_bundle: lint allows
- fault_vfs: fix unused result
- forensics_navigator, fslab: simplify control flow, remove dead code
- impact_graph, isomorphism_proof, lane_selector: refactor match patterns
  and iterator chains
- lock_txn_parity, maintenance_parity: add lint allows for new clippy lints
- metamorphic: cast safety and truncation allows
- oracle: fix unused result binding
- performance_regression_detector, realdb_e2e_logging: minor lint fixes
- release_certificate, replay_harness, replay_triage: clippy compliance
- soak_executor, sql_pipeline_optimization, sql_pipeline_suites: refactors
- sql_semantic_differential, supervision: lint fixes
- t6sv2_checklist: significant simplification of checklist evaluation logic
- tcl_conformance, unit_fixtures, validation_manifest: clippy compliance
- wal_journal_parity: add lint allows
- 12 test files: misc lint allows and minor fixes

Beads:
- issues.jsonl: update bd-2y306.6 status to in_progress with progress notes
  documenting EBR test coverage additions and validation results

+++ crates/fsqlite-harness/src/unit_fixtures.rs
@@ -366,7 +366,7 @@ fn build_empty_page(page_type: PageType, page_size: u32) -> Vec<u8> {
 fn build_single_cell_leaf_page(page_size: u32, _seed: FixtureSeed) -> Vec<u8> {
     let size = page_size as usize;
     let mut page = vec![0u8; size];
-    let header_size: u16 = 8;
+    let header_size: usize = 8;
 
     // Cell content: payload_size(varint) || rowid(varint) || record
     // Record: header_size(varint=2) || serial_type(varint=1 → 8-bit int) || value(1 byte)
@@ -392,7 +392,7 @@ fn build_single_cell_leaf_page(page_size: u32, _seed: FixtureSeed) -> Vec<u8> {
     page[7] = 0; // fragmented free bytes
 
     // Cell pointer array: 2 bytes, big-endian offset to cell content.
-    let ptr_offset = header_size as usize;
+    let ptr_offset = header_size;
     page[ptr_offset] = (co >> 8) as u8;
     page[ptr_offset + 1] = (co & 0xFF) as u8;
 
@@ -407,7 +407,7 @@ fn build_single_cell_leaf_page(page_size: u32, _seed: FixtureSeed) -> Vec<u8> {
 fn build_interior_table_page(page_size: u32, left_child: u32, right_child: u32) -> Vec<u8> {
     let size = page_size as usize;
     let mut page = vec![0u8; size];
-    let header_size: u16 = 12;
+    let header_size: usize = 12;
 
     // Interior table cell: left_child_ptr(4 bytes) || rowid(varint)
     // The divider key is the maximum rowid in the left subtree.
@@ -439,7 +439,7 @@ fn build_interior_table_page(page_size: u32, left_child: u32, right_child: u32) 
     page[11] = right_child as u8;
 
     // Cell pointer array.
-    let ptr_offset = header_size as usize;
+    let ptr_offset = header_size;
     page[ptr_offset] = (co >> 8) as u8;
     page[ptr_offset + 1] = (co & 0xFF) as u8;
 
@@ -454,7 +454,7 @@ fn build_interior_table_page(page_size: u32, left_child: u32, right_child: u32) 
 fn build_multi_cell_leaf_page(page_size: u32, cell_count: u16, seed: FixtureSeed) -> Vec<u8> {
     let size = page_size as usize;
     let mut page = vec![0u8; size];
-    let header_size: u16 = 8;
+    let header_size: usize = 8;
 
     // Build cells from the end of the page backward.
     let mut cell_offsets: Vec<u16> = Vec::new();
@@ -492,7 +492,7 @@ fn build_multi_cell_leaf_page(page_size: u32, cell_count: u16, seed: FixtureSeed
     page[7] = 0;
 
     // Cell pointer array.
-    let ptr_start = header_size as usize;
+    let ptr_start = header_size;
     for (i, &offset) in cell_offsets.iter().enumerate() {
         let pos = ptr_start + i * 2;
         page[pos] = (offset >> 8) as u8;
